MAG-650: Fix website scope legacy login and refactor config fields behavior
